# spring-boot2/src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  description: |
    Apollo配置中心OpenAPI接口文档

    ## 认证方式

    所有API接口都需要通过Authorization header进行身份验证。

    ### 获取Token的方式：

    1. **Portal管理界面获取**：
       - 登录Apollo Portal管理界面
       - 进入"管理员工具" -> "开放平台授权管理"
       - 创建第三方应用，获取Token

    2. **Token格式**：
       ```
       Authorization: token_value
       ```

    3. **Token权限**：
       - Token具有对应应用的读写权限
       - 不同Token可能有不同的环境和命名空间权限
       - 建议为不同用途创建不同的Token

    ---

    ### 使用示例：
    ```bash
    curl -X GET "http://localhost:8070/openapi/v1/apps" \
         -H "Authorization: your_token_here"
    ```
  title: Apollo OpenAPI
  version: 1.0.0
servers:
- url: /
security:
- ApiKeyAuth: []
...
paths:
...
  /openapi/v1/envs/{env}/apps/{appId}/clusters/{clusterName}/namespaces/{namespaceName}/instances:
    get:
      deprecated: false
      description: "GET /openapi/v1/envs/{env}/apps/{appId}/clusters/{clusterName}/namespaces/{namespaceName}/instances"
      parameters:
      - explode: false
        in: path
        name: env
        required: true
        schema:
          type: string
        style: simple
      - description: 应用ID
        explode: false
        in: path
        name: appId
        required: true
        schema:
          type: string
        style: simple
      - description: 集群名称
        explode: false
        in: path
        name: clusterName
        required: true
        schema:
          type: string
        style: simple
      - description: 命名空间名称
        explode: false
        in: path
        name: namespaceName
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                type: integer
          description: 成功返回实例数量
      summary: 获取命名空间下的实例数量
      tags:
      - Instance Management
      x-accepts:
      - application/json
      x-tags:
      - tag: Instance Management
  /openapi/v1/envs/{env}/apps/{appId}/clusters/{clusterName}/namespaces/{namespaceName}/instances_search:
    get:
      deprecated: false
      description: "GET /openapi/v1/envs/{env}/apps/{appId}/clusters/{clusterName}/namespaces/{namespaceName}/instances_search"
      parameters:
      - description: 环境标识
        explode: false
        in: path
        name: env
        required: true
        schema:
          type: string
        style: simple
      - explode: false
        in: path
        name: appId
        required: true
        schema:
          type: string
        style: simple
      - explode: false
        in: path
        name: clusterName
        required: true
        schema:
          type: string
        style: simple
      - explode: false
        in: path
        name: namespaceName
        required: true
        schema:
          type: string
        style: simple
      - explode: true
        in: query
        name: instanceAppId
        required: false
        schema:
          type: string
        style: form
      - description: 页码，从0开始
        example: 0
        explode: true
        in: query
        name: page
        required: true
        schema:
          type: integer
        style: form
      - description: 每页数量
        example: 20
        explode: true
        in: query
        name: size
        required: true
        schema:
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              example:
                page: 0
                size: 0
                total: 0
                content:
                - id: 0
                  appId: ""
                  clusterName: ""
                  dataCenter: ""
                  ip: ""
                  configs:
                  - release:
                      dataChangeCreatedBy: ""
                      dataChangeLastModifiedBy: ""
                      dataChangeCreatedTime: ""
                      dataChangeLastModifiedTime: ""
                      id: 0
                      appId: ""
                      clusterName: ""
                      namespaceName: ""
                      name: ""
                      configurations:
                        ? ""
                        : ""
                      comment: ""
                    releaseDeliveryTime: ""
                    dataChangeLastModifiedTime: ""
                  dataChangeCreatedTime: ""
              schema:
                $ref: "#/components/schemas/OpenPageDTOOpenInstanceDTO"
          description: 成功获取实例列表
      summary: 根据发布版本查询实例（支持分页）
      tags:
      - Instance Management
      x-accepts:
      - application/json
      x-tags:
      - tag: Instance Management
  /openapi/v1/envs/{env}/apps/{appId}/clusters/{clusterName}/namespaces/{namespaceName}/instances_not_in:
    get:
      deprecated: false
      description: "GET /openapi/v1/envs/{env}/apps/{appId}/clusters/{clusterName}/namespaces/{namespaceName}/instances_not_in?excludeReleases=1,2,3"
      parameters:
      - description: 环境标识
        explode: false
        in: path
        name: env
        required: true
        schema:
          type: string
        style: simple
      - description: 应用ID
        explode: false
        in: path
        name: appId
        required: true
        schema:
          type: string
        style: simple
      - description: 集群名称
        explode: false
        in: path
        name: clusterName
        required: true
        schema:
          type: string
        style: simple
      - description: 命名空间名称
        explode: false
        in: path
        name: namespaceName
        required: true
        schema:
          type: string
        style: simple
      - description: 排除的发布ID列表，用逗号分隔
        explode: true
        in: query
        name: excludeReleases
        required: false
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              example:
              - id: 0
                appId: ""
                clusterName: ""
                dataCenter: ""
                ip: ""
                configs:
                - release:
                    dataChangeCreatedBy: ""
                    dataChangeLastModifiedBy: ""
                    dataChangeCreatedTime: ""
                    dataChangeLastModifiedTime: ""
                    id: 0
                    appId: ""
                    clusterName: ""
                    namespaceName: ""
                    name: ""
                    configurations:
                      ? ""
                      : ""
                    comment: ""
                  releaseDeliveryTime: ""
                  dataChangeLastModifiedTime: ""
                dataChangeCreatedTime: ""
              schema:
                items:
                  $ref: "#/components/schemas/OpenInstanceDTO"
                type: array
          description: 成功返回实例列表
      summary: 查询不在指定发布版本中的实例
      tags:
      - Instance Management
      x-accepts:
      - application/json
      x-tags:
      - tag: Instance Management
...
components:
  schemas:
...
    MapString:
      description: 字符串映射对象，用于表示键值对的配置数据结构
      example:
        key: key
      properties:
        key:
          type: string
      type: object
    OpenReleaseDTO:
      description: Apollo发布信息数据传输对象，表示一次配置发布的完整信息
      example:
        dataChangeCreatedTime: dataChangeCreatedTime
        dataChangeLastModifiedBy: dataChangeLastModifiedBy
        configurations:
          key: key
        appId: appId
        clusterName: clusterName
        dataChangeCreatedBy: dataChangeCreatedBy
        name: name
        comment: comment
        id: 0
        dataChangeLastModifiedTime: dataChangeLastModifiedTime
        namespaceName: namespaceName
      properties:
        dataChangeCreatedBy:
          description: 发布创建者用户名，记录是谁创建了这次发布
          type: string
        dataChangeLastModifiedBy:
          description: 发布最后修改者用户名，记录最后一次修改发布信息的用户
          type: string
        dataChangeCreatedTime:
          description: 发布创建时间，ISO 8601格式的时间戳
          type: string
        dataChangeLastModifiedTime:
          description: 发布最后修改时间，ISO 8601格式的时间戳
          type: string
        id:
          description: 发布记录的唯一标识符，系统自动生成
          format: int64
          type: integer
        appId:
          description: 所属应用的唯一标识符
          type: string
        clusterName:
          description: 所属集群的名称
          type: string
        namespaceName:
          description: 所属命名空间的名称
          type: string
        name:
          description: 发布名称，用于标识这次发布的版本或描述
          type: string
        configurations:
          $ref: "#/components/schemas/MapString"
        comment:
          description: 发布备注，描述本次发布的变更内容和目的
          type: string
      type: object
...
    OpenPageDTOOpenInstanceDTO:
      description: 分页实例数据传输对象，用于返回分页查询的实例列表结果
      example:
        total: 1
        size: 6
        page: 0
        content:
        - configs:
          - releaseDeliveryTime: releaseDeliveryTime
            release:
              dataChangeCreatedTime: dataChangeCreatedTime
              dataChangeLastModifiedBy: dataChangeLastModifiedBy
              configurations:
                key: key
              appId: appId
              clusterName: clusterName
              dataChangeCreatedBy: dataChangeCreatedBy
              name: name
              comment: comment
              id: 0
              dataChangeLastModifiedTime: dataChangeLastModifiedTime
              namespaceName: namespaceName
            dataChangeLastModifiedTime: dataChangeLastModifiedTime
          - releaseDeliveryTime: releaseDeliveryTime
            release:
              dataChangeCreatedTime: dataChangeCreatedTime
              dataChangeLastModifiedBy: dataChangeLastModifiedBy
              configurations:
                key: key
              appId: appId
              clusterName: clusterName
              dataChangeCreatedBy: dataChangeCreatedBy
              name: name
              comment: comment
              id: 0
              dataChangeLastModifiedTime: dataChangeLastModifiedTime
              namespaceName: namespaceName
            dataChangeLastModifiedTime: dataChangeLastModifiedTime
          dataChangeCreatedTime: dataChangeCreatedTime
          dataCenter: dataCenter
          appId: appId
          clusterName: clusterName
          ip: ip
          id: 5
        - configs:
          - releaseDeliveryTime: releaseDeliveryTime
            release:
              dataChangeCreatedTime: dataChangeCreatedTime
              dataChangeLastModifiedBy: dataChangeLastModifiedBy
              configurations:
                key: key
              appId: appId
              clusterName: clusterName
              dataChangeCreatedBy: dataChangeCreatedBy
              name: name
              comment: comment
              id: 0
              dataChangeLastModifiedTime: dataChangeLastModifiedTime
              namespaceName: namespaceName
            dataChangeLastModifiedTime: dataChangeLastModifiedTime
          - releaseDeliveryTime: releaseDeliveryTime
            release:
              dataChangeCreatedTime: dataChangeCreatedTime
              dataChangeLastModifiedBy: dataChangeLastModifiedBy
              configurations:
                key: key
              appId: appId
              clusterName: clusterName
              dataChangeCreatedBy: dataChangeCreatedBy
              name: name
              comment: comment
              id: 0
              dataChangeLastModifiedTime: dataChangeLastModifiedTime
              namespaceName: namespaceName
            dataChangeLastModifiedTime: dataChangeLastModifiedTime
          dataChangeCreatedTime: dataChangeCreatedTime
          dataCenter: dataCenter
          appId: appId
          clusterName: clusterName
          ip: ip
          id: 5
      properties:
        page:
          description: 当前页码，从0开始计数
          type: integer
        size:
          description: 每页显示的记录数量
          type: integer
        total:
          description: 总记录数，符合查询条件的实例总数量
          format: int64
          type: integer
        content:
          description: 当前页的实例列表，包含具体的实例数据
          items:
            $ref: "#/components/schemas/OpenInstanceDTO"
          type: array
      type: object
...
    OpenInstanceConfigDTO:
      example:
        releaseDeliveryTime: releaseDeliveryTime
        release:
          dataChangeCreatedTime: dataChangeCreatedTime
          dataChangeLastModifiedBy: dataChangeLastModifiedBy
          configurations:
            key: key
          appId: appId
          clusterName: clusterName
          dataChangeCreatedBy: dataChangeCreatedBy
          name: name
          comment: comment
          id: 0
          dataChangeLastModifiedTime: dataChangeLastModifiedTime
          namespaceName: namespaceName
        dataChangeLastModifiedTime: dataChangeLastModifiedTime
      properties:
        release:
          $ref: "#/components/schemas/OpenReleaseDTO"
        releaseDeliveryTime:
          description: 配置下发时间（ISO 8601）
          type: string
        dataChangeLastModifiedTime:
          description: 最后修改时间（ISO 8601）
          type: string
      type: object
    OpenInstanceDTO:
      example:
        configs:
        - releaseDeliveryTime: releaseDeliveryTime
          release:
            dataChangeCreatedTime: dataChangeCreatedTime
            dataChangeLastModifiedBy: dataChangeLastModifiedBy
            configurations:
              key: key
            appId: appId
            clusterName: clusterName
            dataChangeCreatedBy: dataChangeCreatedBy
            name: name
            comment: comment
            id: 0
            dataChangeLastModifiedTime: dataChangeLastModifiedTime
            namespaceName: namespaceName
          dataChangeLastModifiedTime: dataChangeLastModifiedTime
        - releaseDeliveryTime: releaseDeliveryTime
          release:
            dataChangeCreatedTime: dataChangeCreatedTime
            dataChangeLastModifiedBy: dataChangeLastModifiedBy
            configurations:
              key: key
            appId: appId
            clusterName: clusterName
            dataChangeCreatedBy: dataChangeCreatedBy
            name: name
            comment: comment
            id: 0
            dataChangeLastModifiedTime: dataChangeLastModifiedTime
            namespaceName: namespaceName
          dataChangeLastModifiedTime: dataChangeLastModifiedTime
        dataChangeCreatedTime: dataChangeCreatedTime
        dataCenter: dataCenter
        appId: appId
        clusterName: clusterName
        ip: ip
        id: 5
      properties:
        id:
          description: 实例ID
          format: int64
          type: integer
        appId:
          description: 实例所属应用ID
          type: string
        clusterName:
          description: 实例所属集群名称
          type: string
        dataCenter:
          description: 数据中心/机房标识
          type: string
        ip:
          description: 实例IP地址
          type: string
        configs:
          description: 实例当前生效的配置列表
          items:
            $ref: "#/components/schemas/OpenInstanceConfigDTO"
          type: array
        dataChangeCreatedTime:
          description: 数据创建时间（ISO 8601）
          type: string
      type: object
...
  securitySchemes:
    ApiKeyAuth:
      description: |
        Apollo OpenAPI访问令牌
        **获取方式：**
        1. 登录Apollo Portal管理界面
        2. 进入"管理员工具" -> "开放平台授权管理"
        3. 创建第三方应用，获取Token
        **使用格式：**
        ```
        Authorization: your_token_here
        ```
        **权限说明：**
        - Token具有对应应用的读写权限
        - 不同Token可能有不同的环境和命名空间权限
        - 建议为不同用途创建不同的Token
      in: header
      name: Authorization
      type: apiKey